feat: Add `ChainOpType` to chain integrity warrants
