+++ crates/holochain/src/core/workflow/app_validation_workflow/tests.rs
@@ -1225,6 +1225,7 @@ async fn app_validation_produces_warrants() {
             action_author,
             action: (hash, _),
             validation_type: _,
+            chain_op_type: _,
         }) => {
             assert_eq!(action_author, alice.agent_pubkey());
             assert_eq!(*hash, invalid_action_hash);
