+++ crates/holochain_cascade/src/authority/test.rs
@@ -185,6 +185,7 @@ async fn get_agent_activity() {
             action_author: td.agent.clone(),
             action: (fixt!(ActionHash), fixt!(Signature)),
             validation_type: ValidationType::Sys,
+            chain_op_type: ChainOpType::StoreRecord,
         }),
         fixt!(AgentPubKey),
         td.agent.clone(),
@@ -194,6 +195,7 @@ async fn get_agent_activity() {
             action_author: td.agent.clone(),
             action: (fixt!(ActionHash), fixt!(Signature)),
             validation_type: ValidationType::Sys,
+            chain_op_type: ChainOpType::StoreRecord,
         }),
         fixt!(AgentPubKey),
         td.agent.clone(),
