+++ crates/holochain_state/src/query/live_record.rs
@@ -1,6 +1,6 @@
 use holo_hash::*;
 use holochain_sqlite::rusqlite::named_params;
-use holochain_types::dht_op::ChainOpType;
+use holochain_zome_types::op::ChainOpType;
 use holochain_zome_types::prelude::*;
 use std::fmt::Debug;
