+++ crates/holochain_state/src/query/record_details.rs
@@ -1,7 +1,7 @@
 use holo_hash::*;
 use holochain_sqlite::rusqlite::named_params;
-use holochain_types::dht_op::ChainOpType;
 use holochain_types::prelude::Judged;
+use holochain_zome_types::op::ChainOpType;
 use holochain_zome_types::prelude::*;
 use std::fmt::Debug;
